// apps/web/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ReportStatus = {
  DRAFT: 'DRAFT',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  NEEDS_REVISION: 'NEEDS_REVISION',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const ExpenseType = {
  RECEIPT: 'RECEIPT',
  TRAVEL: 'TRAVEL',
  FOOD: 'FOOD'
} as const

export type ExpenseType = (typeof ExpenseType)[keyof typeof ExpenseType]


export const NotificationPreference = {
  ALL: 'ALL',
  STATUS_CHANGES: 'STATUS_CHANGES',
  NONE: 'NONE'
} as const

export type NotificationPreference = (typeof NotificationPreference)[keyof typeof NotificationPreference]


export const LegalAcceptanceType = {
  CHECKBOX_AND_BUTTON: 'CHECKBOX_AND_BUTTON'
} as const

export type LegalAcceptanceType = (typeof LegalAcceptanceType)[keyof typeof LegalAcceptanceType]